Too quick, was no overflow. Solution right.

+++ cpp/src/problem012.cpp
@@ -39,7 +39,7 @@ public:
         swap(first.t_number, second.t_number);
     }
 
-    long long next() {
+    long next() {
         this->t_number += t_add;
         this->t_add += 1;
         return this->t_number;
@@ -49,17 +49,17 @@ public:
         return this->t_add - 1;
     }
 
-    long long number() {
+    long number() {
         return this->t_number;
     }
 
 private:
     long t_add;
-    long long t_number;
+    long t_number;
 };
 
 /************** Global Vars & Functions *******************/
-void find_divisors(long long tval, std::vector<int> &v) {
+void find_divisors(long tval, std::vector<int> &v) {
     for (int i = 1; i <= tval; ++i) {
         if ((tval % i) == 0) {
             v.push_back(i);
